+++ src/Renderbuffer.cpp
@@ -9,7 +9,7 @@ PyObject * MGLContext_renderbuffer(MGLContext * self, PyObject * args) {
 	int samples;
 
 	const char * dtype;
-	int dtype_size;
+	Py_ssize_t dtype_size;
 
 	int args_ok = PyArg_ParseTuple(
 		args,
