+++ src/Texture3D.cpp
@@ -14,7 +14,7 @@ PyObject * MGLContext_texture3d(MGLContext * self, PyObject * args) {
 	int alignment;
 
 	const char * dtype;
-	int dtype_size;
+	Py_ssize_t dtype_size;
 
 	int args_ok = PyArg_ParseTuple(
 		args,
